refactor(api): :recycle: refactor functions into service 1814
Fixes #1814
